refactor: 이미지에대한 예외처리 구현 및 디렉토리 구조 변경 - 이미지 처리 함수를 process_image.py모듈로 분리한다. - 캣타워를 체크하는 함수를 get_cat_tower.py모듈로 분리한다. - crud클래스를 접근할 데이터베이스에 ㄷ따라 모듈로 분리한다.   crud_mysql.py, crud_redis.py - 쓸데없이 비동기 처리 된 함수를 수정한다. - 모든 엔드포인트가 비동기로 동작하도록 한다. - redis에 대한 예외처리를 추가한다. - 이미지에 대한 예외처리를 추가한다.
